Remove test mode support
When this plugin was originally created there was a separate URL
endpoint for testing Komoju payments. Since then the sandbox URL has
been disabled and determing whether the request is a test or a live
payment is determined by the secret key used. Therefore, having a
separate test setting no longer has any affect, and in fact causes the
payment flow to break, since it's making a request to API that no longer
exists.