Update dev environment to support ngrok tunnelling
Because the new process for using the Hosted Page API uses webhooks, and
those webhooks must have a valid internet-accessible URL to be sent to,
the development WordPress instance needs to be accessible to the
internet. Unfortunately, WordPress doesn't support ngrok tunnelling out
of the box, so some additional config needed to be updated as per
ngrok's docs: https://ngrok.com/docs#wordpress. The extra wordpress
config is being set as per the wordpress image docs:
https://hub.docker.com/_/wordpress/.