Add WP_DEBUG_LOG option to WP config
Because the webhook response is coming out of the regular user driven
flow of the website, we need to enable logging to file to be able to see
any PHP errors that occur processing the webhook request. Specifically
setting the log file to be the /tmp directory because the default
directory is wp-content, which is controlled by root and the WP user
doesn't have permissions to write there.