Add new PHP class to encapsulate the webhook data.
Since the callback_url process is deprecated we need to switch to
using webhooks instead. And to prevent future change to the webhooks
requiring changes throughout the system a new class is being added to
encapsulate the webhook events. This way future changes can be contained
to the new class and don't require the calling code to have to be
updated each time.