Use webhook instead of callback_url
The process for sending a callback_url to Komoju is deprecated and
should no longer be used. The replacement for this is to configure the
Komoju account to send events to a webhook. Since the webhook sends
slightly different data than the callback_url, the receiving had to be
updated to accommodate. While this work was going on I decided to
encapsulate the webhook event data into it's own class, so any future
changes to the webhook data don't need to be updated in multiple places.