Refactor function to use case statement
Refactoring the metaprogramming to use a case statement instead, which
is a bit more straight forward and obvious as to which method is being
called. Using metaprogramming wasn't giving us much benefit as we would
have to create a new method with each new webhook event that needed to
be supported, so we would need to make changes to this code anyway.