Add error handling when parsing the JSON body
If there's an error when parsing the webhook request body, then return a
500 with the error message. Since the Komoju console displays the body
of the response this should give a users/devs a better idea of
the errors and how to fix them.

PHP did introduce an option for the json_decode function to throw an
error if it encounters a problem during parsing, but that was only added
in 7.3.0 and WooCommerce supports 7 and up, so we can't use that here,
despite the fact that would be a cleaner solution.