Remove handling of the payment.failed captured
The payment.failed event is fired when the customer fails to make a
payment from the hosted page, because of an invalid or expired card,
etc. At this point, they are still able to changes the details and
resubmit, so it's fully possible that we could have one or more
payment.failed events come through the system before getting a
payment.captured or payment.authorized event. Because the payment may
succeed after a payment.failed event has been emitted, we don't want
make the order as cancelled.

In this case the simplest answer is to just ignore the payment.failed
event. If the payment is successful then the corresponding event will be
fired and if it isn't then the user will either hit the cancel button or
close the window. In any of these cases knowing about any failures in
between does not matter to the WooCommerce system.