Add store assets to repo
WordPress allows us to add icons and banners to the plugin's store page
by uploading images to the assets/ directory in the SVN repo. Because
having multiple sources of truth becomes a headache, I'm copying across
the assets from the SVN repo to make sure this Git repo has all the
assets.

Also adding a README on how to upload a new version to the store.