Don't send tax as a separate value
The tax param for the hosted page API is an old parameter and going to
be deprecated. Because KOMOJU will eventually become tax agnostic we
will roll the tax value into the transaction[amount] param and set
transaction[tax] to 0. The tax param is still required by the KOMOJU api
at the moment, so it needs to remain.